Async executor configuration
